Rename DataTablePro to DataGrid in Design System
