+++ Source/DesignSystem/molecules/DataGrid/DataGrid.tsx
@@ -7,7 +7,7 @@ import { DataGridPro, DataGridProProps } from '@mui/x-data-grid-pro';
 
 import { Paper } from '@mui/material';
 
-export const DataTablePro = (args: DataGridProProps) =>
+export const DataGrid = (args: DataGridProProps) =>
     <Paper sx={{ width: 1, height: 1, minHeight: 250, boxShadow: 'none' }}>
         <DataGridPro {...args} />
     </Paper>;
